forgot to refactor tests and rename from containers -> components
