+++ src/react-apps/applications/runtime/src/shared/components/altinnAppHeader.tsx
